face_tracking/requirements.txt
ultralytics==8.1.25
opencv-python
numpy==1.24.4
torch==2.1.2
torchvision==0.16.2
scikit-image
filterpy
deep_sort_realtime
deepface
faiss-cpu
